+++ CHANGELOG.md
@@ -9,6 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - White logo for header mobile navigation. 
 - Pagination styles.
 
+## [0.9.5] - 2021-10-05
+### Added
+- ACF resources custom block and styles to suppot it.
+- Search form on header.
+- Search icom (button) on header and fixed it.
+
+## [0.9.4] - 2021-10-05
+### Change
+- Fixed navigation styles
+- Fixed Blog post (from partial) styles. Fluid width.
+- shrinking logo on mobile.
+
 ## [0.9.3] - 2021-19-04
 ### Added
 - Support for posts partials, `archive-blog` and `archive-grid`.